+++ src/billiards/obstacles.jl
@@ -1,6 +1,6 @@
-export Obstacle, Disk, Antidot, RandomDisk,
+export Obstacle, Disk, Antidot, RandomDisk, Wall, Circular,
 InfiniteWall, PeriodicWall, RandomWall, SplitterWall, FiniteWall,
-normalvec, distance, cellsize, Wall, Circular, Semicircle,
+normalvec, distance, cellsize, Semicircle,
 arclength, totallength
 
 #######################################################################################
@@ -137,7 +137,7 @@ Faster than [`FiniteWall`](@ref), meant to be used for convex billiards.
 * `ep::SVector{2,T}` : Ending point of the Wall.
 * `normal::SVector{2,T}` : Normal vector to the wall, pointing to where the
   particle *will come from before a collision* (pointing towards the inside of the
-  billiard table). The size of the vector is irrelevant
+  billiard). The size of the vector is irrelevant
   since it is internally normalized.
 * `name::String` : Name of the obstacle, given for user convenience.
   Defaults to "Wall".
@@ -167,7 +167,7 @@ Slower than [`InfiniteWall`](@ref), meant to be used for non-convex billiards.
 
 Giving a `true` value to the field `isdoor` designates this obstacle to be a `Door`.
 This is used in [`escapetime`](@ref) function. A `Door` is a obstacle of the
-billiard table that the particle can escape from, thus enabling calculations
+billiard that the particle can escape from, thus enabling calculations
 of escape times.
 
 ### Fields:
@@ -175,7 +175,7 @@ of escape times.
 * `ep::SVector{2,T}` : Ending point of the Wall.
 * `normal::SVector{2,T}` : Normal vector to the wall, pointing to where the
   particle *will come from before a collision* (pointing towards the inside of the
-  billiard table). The size of the vector is irrelevant
+  billiard). The size of the vector is irrelevant
   since it is internally normalized.
 * `isdoor::Bool` : Flag of whether this `FiniteWall` instance is a "Door".
 * `name::String` : Name of the obstacle, given for user convenience.
@@ -215,7 +215,7 @@ Wall obstacle imposing (uniformly) random reflection during collision (immutable
 * `ep::SVector{2,T}` : Ending point of the Wall.
 * `normal::SVector{2,T}` : Normal vector to the wall, pointing to where the
   particle *is expected to come from* (pointing towards the inside of the
-  billiard table).
+  billiard).
 * `name::String` : Name of the obstacle, given for user convenience.
   Defaults to "Random wall".
 """
@@ -243,7 +243,7 @@ Wall obstacle that imposes periodic boundary conditions upon collision (immutabl
 * `sp::SVector{2,T}` : Starting point of the Wall.
 * `ep::SVector{2,T}` : Ending point of the Wall.
 * `normal::SVector{2,T}` : Normal vector to the wall, pointing to where the
-  particle *will come from* (to the inside the billiard table).
+  particle *will come from* (to the inside the billiard).
   The size of the vector is **important**!
   This vector is added to a particle's `pos` during collision. Therefore the
   size of the normal vector must be correctly associated with the size of the
@@ -331,7 +331,7 @@ assumed to be very close to the obstacle's boundary).
 #######################################################################################
 
 """
-    function arclength(p::AbstractParticle, o::Obstacle)
+    arclength(p::AbstractParticle, o::Obstacle)
 Returns the boundary coordinate of the particle on the obstacle,
 assuming that the particle position is on the obstacle.
 
@@ -362,9 +362,8 @@ end
 
 
 """
-    function totallength(o::Obstacle)
-Returns the total length of `o`, i.e. the maximum value `arclength(…, o)` can
-return.
+    totallength(o::Obstacle)
+Return the total length of `o`.
 """
 @inline totallength(o::Wall) = norm(o.ep - o.sp)
 @inline totallength(o::Semicircle) = π*o.r
@@ -391,10 +390,10 @@ Return the **signed** distance between particle `p` and obstacle `o`, based on
 of the `Obstacle`. E.g. for a `Disk`, the distance is positive when the particle is
 outside of the disk, negative otherwise.
 
-    distance(p::AbstractParticle, bt::BilliardTable)
+    distance(p::AbstractParticle, bt::Billiard)
 Return minimum `distance(p, obst)` for all `obst` in `bt`.
 If the `distance(p, bt)` is negative this means that the particle is outside
-the billiard table.
+the billiard.
 
 All `distance` functions can also be given a position (vector) instead of a particle.
 """
@@ -463,7 +462,7 @@ end
 ####################################################
 """
     cellsize(bt)
-Return the delimiters `xmin, ymin, xmax, ymax` of the given obstacle/billiard table.
+Return the delimiters `xmin, ymin, xmax, ymax` of the given obstacle/billiard.
 
 Used in `randominside()`, error checking and plotting.
 """