+++ src/poincaresos.jl
@@ -15,13 +15,13 @@ end
 
 
 """
-    arcintervals(bt::BilliardTable)
+    arcintervals(bt::Billiard)
 Generate an array of `SVector`s, with the `i`th `SVector` containing the arc
 length intervals corresponding to the `i`th `Obstacle` in `bt`.
 
 Used by [`poincaresection`](@ref) to compute arc lengths.
 """
-function arcintervals(bt::BilliardTable{T, D}) where {T, D}
+function arcintervals(bt::Billiard{T, D}) where {T, D}
     intervals = Vector{SVector{2,T}}(D)
     current = zero(T)
     for i ∈ 1:D
@@ -36,21 +36,21 @@ end
 
 """
 ```julia
-poincaresection(bt::BilliardTable, t, ps::Vector{<:AbstractParticle})
-poincaresection(bt::BilliardTable, t, n::Int [, ω])
+poincaresection(bt::Billiard, t, ps::Vector{<:AbstractParticle})
+poincaresection(bt::Billiard, t, n::Int [, ω])
 ```
-Compute the poincare section (also called boundary map) of the
-billiard table `bt` by evolving each particle for total amount `t` (either float for
+Compute the Poincaré section (also called boundary map) of the
+billiard `bt` by evolving each particle for total amount `t` (either float for
 time or integer for collision number). See below for the returned values.
 
 If `n::Int` is given instead of `ps`,
-then `n` random particles are produced in the billiard table. If `ω` is also
-given, then the particles are magnetic.
+then `n` random particles are produced in the billiard table using
+[`randominside`](@ref). If `ω` is also given, then the particles are magnetic.
 
-The sorting and measurement direction of the arclengths of the individual obstacles
-is dictated by the `sortorder` field of `bt`, see [`BilliardTable`](@ref) for more.
+The measurement direction of the arclengths of the individual obstacles
+is dictated by `bt`, see [`Billiard`](@ref) for more.
 
-## Returns
+Return
 * the arclengths at the collisions `ξs`
 * the incidence angles at the collisions `φs`
 * obstacle arclength `intervals`
@@ -59,11 +59,9 @@ Both `ξs` and `φs` are vectors of `Vector`.
 The `i` inner vectors correspond to the results of the `i` initial condition/particle.
 
 The `intervals` is a vector of `SVector`. The `i` entry of `intervals` is the
-arclength spanned by the `i` obstacle of the billiard table. The direction
-of the measurement of the arclength is dictated by `bt.sortorder`, see
-[`BilliardTable`](@ref) for more.
+arclength spanned by the `i` obstacle of the billiard table.
 """
-function poincaresection(bt::BilliardTable{T}, t,
+function poincaresection(bt::Billiard{T}, t,
                          ps::Vector{<:AbstractParticle{T}}) where {T}
 
     params = Vector{T}[]
@@ -84,11 +82,7 @@ function poincaresection(bt::BilliardTable{T}, t,
             if typeof(bt[i]) <: PeriodicWall
                 continue # do not write output if collision with with PeriodicWall
             else
-                if !bt.inverted[i]
-                    push!(pparams, arclength(p, bt[i]) + intervals[i][1])
-                else
-                    push!(pparams, intervals[i][2] - arclength(p, bt[i]))
-                end
+                push!(pparams, arclength(p, bt[i]) + intervals[i][1])
                 push!(pangles, reflection_angle(p, bt[i]))
                 # set counter
                 count += increment_counter(t, t_to_write)
@@ -103,8 +97,8 @@ function poincaresection(bt::BilliardTable{T}, t,
     return params, angles, intervals
 end
 
-poincaresection(bt::BilliardTable, t, n::Int) =
+poincaresection(bt::Billiard, t, n::Int) =
     poincaresection(bt, t, [randominside(bt) for i ∈ 1:n])
 
-poincaresection(bt::BilliardTable, t, n::Int, ω::AbstractFloat) =
+poincaresection(bt::Billiard, t, n::Int, ω::AbstractFloat) =
     poincaresection(bt, t, [randominside(bt, ω) for i ∈ 1:n])